requirements.txt
alembic==1.10.3
async-generator==1.10
attrs==23.1.0
autopep8==2.0.2
boto3==1.26.115
botocore==1.29.115
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.6
dulwich==0.21.3
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.2
more-itertools==9.1.0
msgpack==1.0.5
multi-key-dict==2.0.3
outcome==1.2.0
packaging==23.1
pbr==5.11.1
pexpect==4.8.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jenkins==1.8.0
pywin32-ctypes==0.2.0
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
s3transfer==0.6.0
selenium==4.9.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.9
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
trove-classifiers==2023.4.29
typing_extensions==4.5.0
urllib3==1.26.15
utils==1.0.1
virtualenv==20.21.1
webdriver-manager==3.8.6
webencodings==0.5.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
